cleaning up some code, adding comments
